requirements.txt
Django
djangorestframework
django-cors-headers
PILLOW
# tensorflow